Dev: move to Traits for 5.X
